+++ console/src/components/Services/RemoteSchema/Relationships/RemoteSchemaRelationRenderer.tsx
@@ -3,8 +3,12 @@ import { useGetAllRemoteSchemaRelationships } from '@/features/MetadataAPI';
 import { RemoteSchemaRelationshipTable } from '@/features/RelationshipsTable';
 import { Button } from '@/new-components/Button';
 import { RiAddCircleFill } from 'react-icons/ri';
-// eslint-disable-next-line no-restricted-imports
-import { RemoteSchemaToDbForm } from '@/features/RemoteRelationships/RemoteSchemaRelationships/components/RemoteSchemaToDB';
+import {
+  RemoteSchemaToRemoteSchemaForm,
+  RemoteRelOption,
+  RemoteSchemaToDbForm,
+} from '@/features/RemoteRelationships';
+import { IndicatorCard } from '@/new-components/IndicatorCard';
 
 type RemoteSchemaRelationRendererProp = {
   remoteSchemaName: string;
@@ -18,27 +22,46 @@ export const RemoteSchemaRelationRenderer = ({
     isError,
   } = useGetAllRemoteSchemaRelationships();
   const [isFormOpen, setIsFormOpen] = useState(false);
+  const [formState, setFormState] = useState<RemoteRelOption>('remoteSchema');
+
+  if (isError) {
+    return <div>Error in fetching remote schema relationships.</div>;
+  }
+  if (isLoading) {
+    return <div>Loading...</div>;
+  }
 
   return (
     <>
-      {isError ? (
-        <div>Error in fetching remote schema relationships.</div>
-      ) : isLoading ? (
-        <div>Fetching remote schema relationships...</div>
-      ) : (
+      {remoteSchemaRels?.length ? (
         <RemoteSchemaRelationshipTable
-          remoteSchemaRels={remoteSchemaRels ?? []}
+          remoteSchemaRels={remoteSchemaRels}
           showActionCell={false}
           remoteSchema={remoteSchemaName}
         />
+      ) : (
+        <>
+          <IndicatorCard status="info">
+            No remote schema relationships found!
+          </IndicatorCard>
+          <br />
+        </>
       )}
-
       {isFormOpen ? (
-        <RemoteSchemaToDbForm
-          sourceRemoteSchema={remoteSchemaName}
-          closeHandler={() => setIsFormOpen(!isFormOpen)}
-          onSuccess={() => setIsFormOpen(false)}
-        />
+        formState === 'remoteSchema' ? (
+          <RemoteSchemaToRemoteSchemaForm
+            sourceRemoteSchema={remoteSchemaName}
+            closeHandler={() => setIsFormOpen(!isFormOpen)}
+            relModeHandler={setFormState}
+          />
+        ) : (
+          <RemoteSchemaToDbForm
+            sourceRemoteSchema={remoteSchemaName}
+            closeHandler={() => setIsFormOpen(!isFormOpen)}
+            onSuccess={() => setIsFormOpen(false)}
+            relModeHandler={setFormState}
+          />
+        )
       ) : (
         <Button
           icon={<RiAddCircleFill />}