+++ console/src/features/RemoteRelationships/RemoteSchemaRelationships/components/RemoteSchemaToDB/FormElements.tsx
@@ -1,15 +1,18 @@
+import React, { useEffect, useState } from 'react';
+import { useFormContext } from 'react-hook-form';
 import { useRemoteSchema } from '@/features/MetadataAPI';
 // eslint-disable-next-line no-restricted-imports
 import { useTableColumns } from '@/features/SqlQueries/hooks/useTableColumns';
 import { InputField, Select } from '@/new-components/Form';
 import { MapSelector } from '@/new-components/MapSelector';
-import React, { useEffect, useState } from 'react';
-import { useFormContext } from 'react-hook-form';
-import { FaLink } from 'react-icons/fa';
+import {
+  LinkBlockHorizontal,
+  LinkBlockVertical,
+} from '@/new-components/LinkBlock';
 import { RemoteDatabaseWidget } from '../RemoteDatabaseWidget';
 import { RsSourceTypeSelector } from '../RsSourceTypeSelector';
 import { Schema } from './schema';
-import { getTypesFromIntrospection } from './utils';
+import { getTypesFromIntrospection } from '../../utils';
 
 export const FormElements = ({
   sourceRemoteSchema,
@@ -65,27 +68,7 @@ export const FormElements = ({
 
   return (
     <>
-      <hr className="mb-md border-gray-300" />
-
-      {/* relationship meta */}
-      <div className="mb-md">
-        <div className="grid gap-sm grid-cols-1 sm:grid-cols-2">
-          <div className="grid gap-sm grid-cols-1 sm:grid-cols-2">
-            <div className="bg-white shadow-sm rounded p-md border border-gray-300">
-              <p className="flex items-center font-semibold text-muted">
-                <label className="cursor-pointer ml-sm font-semibold">
-                  Remote Schema Relationship
-                </label>
-              </p>
-              <p className="text-muted pl-6">
-                Relationship from this remote schema to another
-                database&nbsp;schema.
-              </p>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="w-full sm:w-6/12 mb-md">
+      <div className="w-full sm:w-6/12 my-md">
         <div className="mb-md">
           <InputField
             name="relationshipName"
@@ -123,20 +106,7 @@ export const FormElements = ({
           />
         </div>
 
-        {/* horizontal connector line */}
-        <div className="col-span-2 flex relative items-center justify-center w-full py-md">
-          <div
-            className="flex z-10 items-center justify-center border border-gray-300 bg-white"
-            style={{
-              height: '32px',
-              width: '32px',
-              borderRadius: '100px',
-            }}
-          >
-            <FaLink />
-          </div>
-          <div className="absolute w-full border-b border-gray-300" />
-        </div>
+        <LinkBlockHorizontal />
 
         <div className="col-span-5">
           <RemoteDatabaseWidget />
@@ -144,16 +114,7 @@ export const FormElements = ({
       </div>
 
       {/* vertical connector line */}
-      <div className="flex items-center w-full px-8">
-        <div className="relative flex items-center justify-center h-20">
-          <div className="absolute border border-l border-gray-300 h-20" />
-          <div className="absolute border border-gray-300 flex items-center justify-center rounded-full h-10 w-10 bg-white">
-            <FaLink />
-          </div>
-        </div>
-        <p className="m-0 px-8 font-semibold">Type Mapped To</p>
-      </div>
-
+      <LinkBlockVertical title="Type Mapped To" />
       <MapSelector
         types={
           remoteSchemaTypes.find(x => x.typeName === RSTypeName)?.fields ?? []