+++ console/src/features/RemoteRelationships/index.ts
@@ -1,0 +1,3 @@
+export * from './RemoteSchemaRelationships/components/RemoteSchemaToRemoteSchemaForm/RemoteSchemaToRemoteSchemaForm';
+export * from './RemoteSchemaRelationships/components/RemoteSchemaToRemoteSchemaForm/RelationshipTypeCardRadioGroup';
+export * from './RemoteSchemaRelationships/components/RemoteSchemaToDB';