+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - console: add support for setting aggregation query permissions for ms sql server
 - console: add RS-to-DB (only postgres & citus) relationships feature to remote schemas tab
 - console: remove need for clicking the Modify btn before editing a remote schema (#1193, #8262)
+- console: integrate the RS-to-RS form into Remote Schema Relationship tab
 - cli: fix remote schema metadata formatting issues (#7608)
 - cli: fix query collections metadata formatting issues (#7616)
 - docs: support for `graphql-ws` is considered GA